requirements.txt
langchain
langchain-google-genai
python-dotenv
openai
requests
streamlit

